english zap bot/requirements.txt
python-telegram-bot[job-queue]>=22.0,<23
httpx>=0.27,<1
python-dotenv>=1.0,<2
edge-tts>=7.0,<8
fastapi>=0.115,<1
uvicorn>=0.30,<1
python-multipart>=0.0.9,<1
